- added liked movies button on frontend - added backend api for liked movies

+++ backend-kotlin/src/main/kotlin/com/vidbox/backend/repos/MovieInfoTopRatedRepository.kt
@@ -6,7 +6,8 @@ import org.springframework.data.domain.Pageable
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
 
-interface MovieInfoTopRatedRepository : JpaRepository<MovieInfoTopRated, Long> {
+interface MovieInfoTopRatedRepository : JpaRepository<MovieInfoTopRated, Int> {
+
     //@Query("SELECT m FROM MovieInfoTopRated m WHERE lower(m.title) LIKE lower(concat('%', :query, '%'))")
     @Query(
         value = "SELECT * FROM movie_infos_top_rated m WHERE to_tsvector('english', m.title) @@ to_tsquery('english', :query || ':*')",
@@ -14,4 +15,5 @@ interface MovieInfoTopRatedRepository : JpaRepository<MovieInfoTopRated, Long> {
         nativeQuery = true
     )
     fun findByTitleContains(query: String, pageable: Pageable): Page<MovieInfoTopRated>
+
 }
